Move each page to be in control of the footer + header
