+++ packages/frontend/src/app.tsx
@@ -1,10 +1,6 @@
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faPlus } from "@fortawesome/free-solid-svg-icons";
-import { Route, Routes, type To, useNavigate } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import { useEffect, useRef, useState } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
-import { Page } from "./components/page/index.js";
-import { Header } from "./components/page/header.js";
 import { TestPage } from "./pages/test.js";
 import { AnotherTestPage } from "./pages/another_test.js";
 import { API, type AuthResult, handleApi } from "./state/auth.js";
@@ -15,21 +11,6 @@ export function App() {
   const [result, setResult] = useState({ type: "loading" } as AuthResult);
   const hasSetApi = useRef(false);
   const { isAuthenticated } = useAuth0();
-  const navigate = useNavigate();
-  const viewNavigate = (newRoute: To) => {
-    // Navigate to the new route
-    if (
-      !(document as unknown as { startViewTransition?: unknown })
-        .startViewTransition
-    ) {
-      navigate(newRoute);
-    } else {
-      return document.startViewTransition(() => {
-        navigate(newRoute);
-      });
-    }
-  };
-
   const api = useAPIClient(isAuthenticated);
 
   useEffect(() => {
@@ -54,37 +35,10 @@ export function App() {
 
   return ok.authenticated ? (
     <API.Provider value={ok.client}>
-      <Page>
-        <Page.Header>
-          <Header.Back />
-          Whatever goes in here, stays in here.
-          <Header.Right>
-            <FontAwesomeIcon icon={faPlus} />
-          </Header.Right>
-        </Page.Header>
-        <Page.Body style={{ viewTransitionName: "body-content" }}>
-          <Routes>
-            <Route index element={<TestPage />} />
-            <Route path="/another" element={<AnotherTestPage />} />
-          </Routes>
-        </Page.Body>
-        <Page.Footer>
-          <button
-            onClick={() => {
-              viewNavigate("/");
-            }}
-          >
-            First Test page
-          </button>
-          <button
-            onClick={() => {
-              viewNavigate("/another");
-            }}
-          >
-            Another page.
-          </button>
-        </Page.Footer>
-      </Page>
+      <Routes>
+        <Route index element={<TestPage />} />
+        <Route path="/another" element={<AnotherTestPage />} />
+      </Routes>
     </API.Provider>
   ) : (
     <div className="flex min-h-screen px-3 items-center place-content-center">
